app/requirements.txt
flask==2.0.3
werkzeug==2.0.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
prometheus-client==0.12.0
flasgger